Fix issue #1143: move fake_repo stuff to a pytest fixture
